Proofreading and edits
Line 93-34 - I think it is a little unclear that you are using "begin" and "end" as both the starting and ending points of the document AND as commands to mark those. Maybe add some language highlighting the fact that these are commands to help reinforce the "think like a programmer" mindset?

Lines 303 - 352 - I'm assuming you have to have an internet connection to import packages? If so this should be stated.

Lines 392-394 - Are those the only file types to be created that can be deleted? For a beginner who might not be super familiar knowing what can be deleted can be very nervewracking. If they delete the wrong thing can they just recompile? Are there certain things they should be absolutely sure not to delete? Is deleting them necessary, recommended, or just something they can do if they want? Is there a resource they can be directed to that will help them make this determination?