Proofreading and Editing.
Section title does not match the .md file name. Might be worth changing to be consistent?

Line 006 - There is only one question. Should that be singular? 

Lines 54-75 - Why is "This line of code" indented in the second image and not in the first?

Lines 71-73 - Should this use text instead of line?

Lines 298-344 - Some colorblind participants might not be able to pick up the red. Maybe examples using blue or yellow or gray as well to make the examples more accessible.